Refactored TestContainerConstants and updated Readme
