--- modules/engines/engine-hdb/src/test/java/com/sap/xsk/hdb/ds/test/itest/hdbsequence/utils/TestContainerConstants.java
@@ -1,10 +1,0 @@
-package com.sap.xsk.hdb.ds.test.itest.hdbsequence.utils;
-
-public class TestContainerConstants {
-
-  private TestContainerConstants() {
-  }
-
-  public static final String POSTGRESQL_DOCKER_IMAGE = "postgres:alpine";
-  public static final String POSTGRESQL_DOCKER_NETWORK_ALIAS = "postgres";
-}
